Initial working version of area alerts (#1000)
All the work on district alerts so far, with all basic features working for creating area alerts and managing subscriptions.

Updates all the auth-provider APIs for the new endpoints added for districts, plus some reorganization I did to hopefully make things a bit more clear with paths/naming and align better with the new duel building/districts setup. This includes all the changes to the jfauthprovider intermediate API on the django side, and the AuthClient and UserContext methods on the frontend.

Also adds a new /district page, and the dropdown + map area selection system. This uses static a single json file with all the area types to populate the first dropdown, then with each selection of a type it calls a new wow api to get geojson data for the map and that is also used to populate the dropdown for individual areas of that type. (more details below)

This also updates the existing Login component so that it can work similarly to building alerts where you can be directed there when trying to subscribe to district alerts when not logged in. It will prompt you to register/login, then at the end automatically subscribe you to your selections and shows you a page confirming your subscription with links to email settings.

This also fixes a bug we have encountered before where after you're navigated to a new page after logging in it doesn't have the user data right away. This was because of the same issue we'd seen before of checking for a missing user !user which only works if it's completely undefined, but in some cases it's there but empty (eg. {email: undefined, ...} but the !user on that returns false still. What we need to do is !user?.email, this change was made in AuthCleint and that solved the problem. Now we can navigate users directly to email settings after the complete the district signup via login/register flow.

I was also running into a bug where after subscribing to a district and returning to account setting, the newest district wasn't included in the list. It turns out this was because we were trying to limit api calls to fetch user in AuthClient by updating a var with the user and only fetching new if it's missing - but this means it doesn't get new subscriptions. So I removed that logic, and I don't think it'll be a problem to make an extra request once in a while, since we are already keeping user in UserContext so re-fetching shouldn't be super frequent.

This also adds SQL to create wow_districts_geom a table with a row for each district from all types, with the fields we use for storing area selection in the auth-provider db, as well as the geometry. And a second table wow_districts_geojson that takes each type of districts and creates two geojson feature collections - one for shapes and one for points to place labels. This geojson table is used for a new wow data api endpoint to get the map data when a district type is selected from the dropdown menu. We'll only need to create these ables once, and we don't really have a pattern for that yet - all other tables are always part of a recurring job, so we'll eventually need to make a plan for that.

Currently we are using a custom mapbox style, but it's not compatible with our very outdated version of mapboxgl, so it raises a bunch of console errors but still works. Eventually I want to recreate the non-labeled map style in an old compatible version, it will just be a tedious process of updating colors from a copy of our other wow map, so waiting to do this later.

Currently we have a simple side-by-side map/dropdown layout, but will update soon once the designs are finalized.

This also updates the unsubscribe/manage subscriptions via email page. We no longer have a separate version of the page for "unsubscribe all", only the multipurpose "manage subscriptions". It now includes building and area alert subscriptions, with buttons to unsubscribe from individual buildings/areas and to unsubscribe from all. Still waiting on the final design of the area alerts subscription field (how to display the area components), but that will be updated separately later.

All of this work lines up with this companion PR on auth-provider: JustFixNYC/auth-provider#64