Area Alerts SQL queries & API endpoints (#1005)
* add jfauthprovider api views, update authClient, add basic district page

* add districts to email unsubscribe page for testing

* add district selection dropdowns

* update for auth building api changes

* add district selection page

* trigger branch deploy

* lingui

* move subscription list component, fix clear area selection

* redirect to email settings after district selection

* nest area components on settings page

* remove login requirement for district page, save doesn't connect to db

* update copy on account page

* add errors to district selection page buttons

* add district sign up error/success states

* add flow for: district -> login/register -> subscribe

* lingui

* redirect to home after district subscribe

* remove "add to email" button and updating map link

* remove unused styles after removing "add to email" button

* first draft of district map for area selection

* lingui

* add sql for generating options for district selection dropdowns

* lingui

* fix bug that removes user from context after navigate

* move map into original district page, add linked state

* dont reload map on district type change

* cut map page, move map to component, add paint styles

* fix account district labels

* add zoom controls, fix default zoom/center

* add districts geojson sql/api

* use wow api for district map data

* change district type validator

* typo

* fix col name typo

* add custom map style

* add wow_districts_geom table

* typo

* update geojson creation for new table structure

* space in area names

* add map loading placeholder, add zips, reorder map/dropdowns

* remove old links from district page

* fix lingering map area hover outline

* always re-fetch user from authclient.fetchuser

* reset map style to default labeled one (will fix later)

* add link to add additional area in settings

* disable map rotation

* delete old files, add maxbounds, remove unused packages

* add back mapboxgl types as dev dep

* login success message after district alerts subscribe

* initial side-by-side map selection layout

* update manage subscriptions via email for districts, no separate unsubscribe all page

* full screen district selection

* dropdown space fix

* swap map and sidepane order

* remove debug print

* full screen map and overlay widget on desktop

* lingui

* add create indicators table, email district alert api

* fix widget height w/ dropdowns open

* update map styles

* add mobile version

* format and translate area labels for chips and settings

* add padding for dropdown, increase zoom control size

* zip default map type, sort area labels in dropdown

* reduce zoom control size

* mobile: map selection scrolls down to save, dropdown selection does not

* .env

* get top 5 buildings and portfolios

* remove unused import

* Delete client/.env

* Revert "Delete client/.env"

This reverts commit 424b97bf8855f2f09ade7837c987695d84e2e761.

* remove credential from .env but restore file

* limit top 5 sql to 5

* add sales and vacate order api

* fix all ranks = 1 from a null normalized column

* add litigations

* fix issue where value of 0 should not be ranked, add hpd_link

* run black

* fix Any type hint lint error

* add complaintnumber to dob vacate order

* fix portfolio_size misnomer to area_bbls in alert query

* add variables from vacate order tables, formatted hpd links

* rename pluto_latest_district nta col

* updates for k8s wow job (#1009)

some changes to make sure everything is set up for running the wow job via k8s-loader for the new area alerts work.

We depend on the newly added pluto_latest_districts_25a nycdb dataset, which is a new kind of nycdb dataset that has no source data but is just a sql script to create a table from other nycdb datasets (pluto_latest and boundaries_25a). We're using the pinned version so that the boundaries don't change overtime without us knowing.

The new setup in nycdb to handle this new dataset with dependencies and no source data is not yet finalized, so we're using a commit hash for an in-progress branch that works for our needs now but still needs some improvements to handle the dependency management and empty files and schema sections of the dataset yml definition.

nycdb/nycdb#386

Along with this change here we are also updating the k8s-loader to add the new dataset:
JustFixNYC/nycdb-k8s-loader#199
And this PR that updates to handle the new pre/post sql scripts for wow:
JustFixNYC/nycdb-k8s-loader#201

Since we now have dependencies for nycdb spatial datasets we needed to update the docker image used in our CI tests so that it has access to shp2pgsql because for some reason circle ci's postgis image doesn't include that. So I added a new custom docker image that just builds on the one we were previously using and installs the packages required to import the shapefiles. This is now published to our dockerhub, and we shouldn't have to make any changes to this going forward - so I just did the publish manually.

I spun up a clone of our DB and ran the updated k8s-loader wow job on it and everything worked.

* remove unused vars

* run black

* update sqls to use wow_indicators table only, fix dob complaints subqueries

* sale data interval change to week

* add PAD as wow nycdb dependency

* use 25a tables for obundaries, cast literals to text

---------

Co-authored-by: Maxwell Austensen <m.austensen@gmail.com>
Co-authored-by: Maxwell Austensen <maxwell@justfix.org>