Area alert API for sales and litigations for month relative to updated date (#1015)
For area alerts emails we now want to include the most recent month of data, based on when the dataset was last updated rather than the date we are sending the email. So this edits the api queries to use the table with datasets' last updated dates to create the date filer.

Companion PR on auth-provider for email: JustFixNYC/auth-provider#85