update dev dockerfile for latest geosupport (#1027)
A new version of Geosupport has been released, 25a, and this updates the local development dockerfile for the new download file. It's good to keep this updated regularly so that it keeps pace with the version of PLUTO we get in pluto_latest.

The important place for this to be updated is on nycdb-k8s-loader, since that dockerfile is actually used for the cron jobs that run the WOW job that uses geosupport to geocode HPD registration contact business addresses. see JustFixNYC/nycdb-k8s-loader#207