update bbls of cofos with no longer existing bbls (#1028)
We got some feedback from a user on GoodCauseNYC about a case where a 2010 CofO that appears on the DOB site was not being found on GoodCauseNYC and so gave an incorrect "covered" result. We discovered that this is because for some older CofO records the BBL from the time no longer exists (since they change when lots of split/merged) but in many cases using the BIN you can join with PAD to get the current BBL that the bin is associated with now.

Can we confirmed that this works for the case they found: SELECT * FROM wow_indicators WHERE bbl = '4005660136';